requirements.txt
Flask==2.3.3
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0